+++ src/FeatureSwitchOptions.php
@@ -1,9 +1,0 @@
-<?php
-
-namespace Drupal\feature_switches;
-
-class FeatureSwitchOptions {
-
-  const ALLOW_UNREADY_LIVE = 1;
-
-}