+++ src/OperatorOptions.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Drupal\feature_switches;
+
+class OperatorOptions {
+
+  const REQUIRE_READY_LIVE = 1;
+
+}