+++ tests/Unit/FeatureSwitchesTest.php
@@ -3,7 +3,7 @@
 namespace Drupal\Tests\feature_switches;
 
 use Drupal\feature_switches\Feature;
-use Drupal\feature_switches\FeatureSwitchOptions;
+use Drupal\feature_switches\OperatorOptions;
 use Drupal\feature_switches\Operator;
 use Drupal\feature_switches\FeatureSwitches;
 use PHPUnit\Framework\TestCase;
@@ -25,9 +25,9 @@ class FeatureSwitchesTest extends TestCase {
 
   public function testSetOptionsSetsOnExistingGlobalOperator() {
     $switches = FeatureSwitches::getOperator();
-    $this->assertSame($switches, FeatureSwitches::setOptions(FeatureSwitchOptions::ALLOW_UNREADY_LIVE));
+    $this->assertSame($switches, FeatureSwitches::setOptions(OperatorOptions::REQUIRE_READY_LIVE));
     $this->assertSame($switches, FeatureSwitches::getOperator());
-    $this->assertSame(1, $switches->getOptions() & FeatureSwitchOptions::ALLOW_UNREADY_LIVE);
+    $this->assertSame(1, $switches->getOptions() & OperatorOptions::REQUIRE_READY_LIVE);
   }
 
   public function testSetIsLiveOnNonExistentHasNoEffect() {