+++ knowledge/pages/general/CHANGELOG.md
@@ -14,6 +14,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * None
 
+## [0.0.5] - 2023-10-13
+
+### Changed
+
+- `FeatureSwitchOptions::ALLOW_UNREADY_LIVE` -> `OperatorOptions::REQUIRE_READY_LIVE` and inverted logic.
+- The default behavior is now to allow unready to be live. You must set the option for the exception to throw.
+
 ## [0.0.4] - 2023-10-13
 
 ### Added